Move Deprecation class to Vanilla\Utility namespace
